+++ test/end2end/http_data.dart
@@ -7,7 +7,7 @@ import 'package:test/test.dart';
 import 'shared.dart';
 
 final data = new E2EData('http', '0.11.3+13', _data, {
-  'pkg': 'everywhere',
+  'pkg': 'undefined',
   'package:http/browser_client.dart': 'web',
   'package:http/http.dart': 'everywhere',
   'package:http/testing.dart': 'flutter, server'