+++ test/end2end/pub_server_data.dart
@@ -7,7 +7,7 @@ import 'package:test/test.dart';
 import 'shared.dart';
 
 final data = new E2EData('pub_server', '0.1.1+3', _data, {
-  'pkg': 'everywhere',
+  'pkg': 'undefined',
   'package:pub_server/repository.dart': 'everywhere',
   'package:pub_server/shelf_pubserver.dart': 'everywhere'
 });